chore: remove unused tokenlistUpdater flow
