+++ src/components/Popups/PopupItem.tsx
@@ -5,7 +5,6 @@ import { useSpring } from 'react-spring'
 import styled, { useTheme } from 'styled-components'
 import { PopupContent } from '../../state/application/types'
 import { useRemovePopup } from '../../state/application/hooks'
-import ListUpdatePopup from './ListUpdatePopup'
 import TransactionPopup from './TransactionPopup'
 
 export const StyledClose = styled(X)`
@@ -79,13 +78,6 @@ export default function PopupItem({
             txn: { hash, success, summary },
         } = content
         popupContent = <TransactionPopup hash={hash} success={success} summary={summary} />
-    } else if ('listUpdate' in content) {
-        const {
-            listUpdate: { listUrl, oldList, newList, auto },
-        } = content
-        popupContent = (
-            <ListUpdatePopup popKey={popKey} listUrl={listUrl} oldList={oldList} newList={newList} auto={auto} />
-        )
     }
 
     const faderStyle = useSpring({
