+++ src/state/index.ts
@@ -2,7 +2,6 @@ import { configureStore, getDefaultMiddleware } from '@reduxjs/toolkit'
 import { load, save } from 'redux-localstorage-simple'
 import application from './application/reducer'
 import { updateVersion } from './global/actions'
-import lists from './lists/reducer'
 import multicall from './multicall/reducer'
 import swap from './swap/reducer'
 import transactions from './transactions/reducer'
@@ -17,7 +16,6 @@ const store = configureStore({
         transactions,
         swap,
         multicall,
-        lists,
     },
     middleware: [...getDefaultMiddleware({ thunk: false, immutableCheck: false }), save({ states: PERSISTED_KEYS })],
     preloadedState: load({ states: PERSISTED_KEYS }),
